Creating Backend Entities
Article and basis entity created also added a base entity to merge in for auditing purposes later though dont want this to effect my progress as I am a week behind on timeline. Also decided not to use the basis id as a foreign key in article logically it should be but this was a bit of a learning curve. This would only create issues as I add id later since I will need to up date in front end and backend instead of just backend but I think it will save me time for now.